+++ tests/files/no-unused-modules/typescript/file-ts-a.ts
@@ -1,3 +1,8 @@
 import {b} from './file-ts-b';
+import {c} from './file-ts-c';
+import {d} from './file-ts-d';
+import {e} from './file-ts-e';
 
-export const a = b + 1;
+export const a = b + 1 + e.f;
+export const a2: c = {};
+export const a3: d = {};